SimCodeToLLVM Phase 4.4: emit per-equation IR body
Lower SES_SIMPLE_ASSIGN equations into LLVM IR by walking the
DAE.Exp tree (RCONST, ICONST, CREF, UNARY UMINUS, BINARY +/-/*//)
and emitting realVars loads/stores through the C-ABI shims added
in Phase 4.3.

The HelloWorldSim model with der(x) = -x now produces:

  %t_0 = call double @omc_jit_get_real_var(ptr %data, i64 0)
  %neg = fneg double %t_0
  call void @omc_jit_set_real_var(ptr %data, i64 2, double %neg)
  ret i64 0

i.e. read x from slot 0, negate, write der(x) to slot 2 (which is
the absolute slot for the first derivative in a 2-state-or-fewer
layout).

The lowering avoids matchcontinue and exception based control
flow: emitExp propagates failure via a Boolean ok return, and the
nested-match-in-then construct that confused codegen was
factored into emitCrefRead.

+++ OMCompiler/Compiler/runtime/llvm_gen_wrappers.c
@@ -1240,12 +1240,16 @@ modelica_integer iMod(modelica_integer x, modelica_integer y)
  * the caller is responsible for translating that into the absolute
  * slot offset passed here.
  */
-double omc_jit_get_real_var(DATA *data, int slot)
+/* Use int64_t for the slot rather than int so the C ABI matches the
+ * LLVM IR the JIT emits -- the EXT_LLVM MODELICA_INTEGER constant
+ * resolves to i64 on the architectures we care about, and we want the
+ * helper's parameter to match without a per-call truncation. */
+double omc_jit_get_real_var(DATA *data, int64_t slot)
 {
   return data->localData[0]->realVars[slot];
 }
 
-void omc_jit_set_real_var(DATA *data, int slot, double value)
+void omc_jit_set_real_var(DATA *data, int64_t slot, double value)
 {
   data->localData[0]->realVars[slot] = value;
 }
